requirements.txt
yt-dlp
discord.py[voice]
dotenv
